Case_Study_5/requirement.txt
numpy
pandas
scikit-learn
Langchain
Langchain-Core
python-dotenv
google-generativeai